Skill_prepdocs/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
jsonschema

#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
aiohttp==3.9.1
    # via -r requirements.in
aiosignal==1.3.1
    # via aiohttp
annotated-types==0.6.0
    # via pydantic
anyio==3.7.1
    # via
    #   httpx
    #   openai
attrs==23.1.0
    # via aiohttp
azure-ai-formrecognizer==3.3.2
    # via -r requirements.in
azure-common==1.1.28
    # via
    #   azure-ai-formrecognizer
    #   azure-keyvault-secrets
    #   azure-search-documents
azure-core==1.29.5
    # via
    #   azure-ai-formrecognizer
    #   azure-identity
    #   azure-keyvault-secrets
    #   azure-search-documents
    #   azure-storage-blob
    #   azure-storage-file-datalake
    #   msrest
azure-identity==1.15.0
    # via -r requirements.in
azure-keyvault-secrets==4.7.0
    # via -r requirements.in
azure-search-documents==11.4.0b11
    # via -r requirements.in
azure-storage-blob==12.19.0
    # via
    #   -r requirements.in
    #   azure-storage-file-datalake
azure-storage-file-datalake==12.14.0
    # via -r requirements.in
certifi==2023.11.17
    # via
    #   httpcore
    #   httpx
    #   msrest
    #   requests
cffi==1.16.0
    # via cryptography
charset-normalizer==3.3.2
    # via requests
colorama==0.4.6
    # via tqdm
cryptography==41.0.7
    # via
    #   azure-identity
    #   azure-storage-blob
    #   msal
    #   pyjwt
distro==1.8.0
    # via openai
frozenlist==1.4.0
    # via
    #   aiohttp
    #   aiosignal
h11==0.14.0
    # via httpcore
httpcore==1.0.2
    # via httpx
httpx==0.25.2
    # via openai
idna==3.6
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
isodate==0.6.1
    # via
    #   azure-keyvault-secrets
    #   azure-search-documents
    #   azure-storage-blob
    #   azure-storage-file-datalake
    #   msrest
msal==1.25.0
    # via
    #   azure-identity
    #   msal-extensions
msal-extensions==1.0.0
    # via azure-identity
msrest==0.7.1
    # via azure-ai-formrecognizer
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
numpy==1.26.2
    # via
    #   openai
    #   pandas
    #   pandas-stubs
oauthlib==3.2.2
    # via requests-oauthlib
openai[datalib]==1.3.7
    # via -r requirements.in
pandas==2.1.3
    # via openai
pandas-stubs==2.1.1.230928
    # via openai
pillow==10.1.0
    # via -r requirements.in
portalocker==2.8.2
    # via msal-extensions
pycparser==2.21
    # via cffi
pydantic==2.5.2
    # via openai
pydantic-core==2.14.5
    # via pydantic
pyjwt[crypto]==2.8.0
    # via msal
pymupdf==1.23.7
    # via -r requirements.in
pymupdfb==1.23.7
    # via pymupdf
pypdf==3.17.1
    # via -r requirements.in
python-dateutil==2.8.2
    # via pandas
pytz==2023.3.post1
    # via pandas
regex==2023.10.3
    # via tiktoken
requests==2.31.0
    # via
    #   azure-core
    #   msal
    #   msrest
    #   requests-oauthlib
    #   tiktoken
requests-oauthlib==1.3.1
    # via msrest
six==1.16.0
    # via
    #   azure-core
    #   isodate
    #   python-dateutil
sniffio==1.3.0
    # via
    #   anyio
    #   httpx
    #   openai
tenacity==8.2.3
    # via -r requirements.in
tiktoken==0.5.2
    # via -r requirements.in
tqdm==4.66.1
    # via openai
types-pillow==10.1.0.2
    # via -r requirements.in
types-pytz==2023.3.1.1
    # via pandas-stubs
typing-extensions==4.8.0
    # via
    #   azure-ai-formrecognizer
    #   azure-core
    #   azure-keyvault-secrets
    #   azure-storage-blob
    #   azure-storage-file-datalake
    #   openai
    #   pydantic
    #   pydantic-core
tzdata==2023.3
    # via pandas
urllib3==2.1.0
    # via requests
yarl==1.9.3
    # via aiohttp

python-dotenv==0.19.2
    # via -r requirements.in


